backend/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==3.2.13
django-allauth==0.51.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-webpack-loader==1.6.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
oauthlib==3.2.0
Pillow==9.1.1
psycopg==3.0.15
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.2.0
urllib3==1.26.9
whitenoise==6.2.0
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==3.2.13
django-allauth==0.51.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-webpack-loader==1.6.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
oauthlib==3.2.0
Pillow==9.1.1
psycopg==3.0.15
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.2.0
urllib3==1.26.9
whitenoise==6.2.0
